requirements.txt
accelerate>=0.2
matplotlib 
h5py 
scikit-learn
nltk>=3
numpy>=1.2
openpyxl==3.1.2
packaging==23.1
pandas
parmap==1.7.0
Pillow
salesforce-lavis==1.0.2
torch==1.10.0
torchvision
tqdm
transformers
omegaconf 
iopath
decord
webdataset
opencv-python-headless==4.5.5.64 
fairscale==0.4.4
einops>=0.4.1
spacy
pycocoevalcap
gradio